Relacion entre las camaras/incidencias y source corregido

+++ src/main/java/com/trafikapp/reto/trafikapp/modelo/SourceRepositorio.java
@@ -3,6 +3,6 @@ package com.trafikapp.reto.trafikapp.modelo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SourceRepositorio extends JpaRepository<Sources, Long> {
+public interface SourceRepositorio extends JpaRepository<Sources, Integer> {
 
 }
